+++ apps/src/javalab/javalabRedux.js
@@ -1,70 +1,28 @@
-import UserPreferences from '../lib/util/UserPreferences';
-import {DisplayTheme} from './DisplayTheme';
-import {
-  DEFAULT_FONT_SIZE_PX,
-  FONT_SIZE_INCREMENT_PX,
-  MAX_FONT_SIZE_PX,
-  MIN_FONT_SIZE_PX
-} from './editorThemes';
-
-const COLOR_PREFERENCE_UPDATED = 'javalab/COLOR_PREFERENCE_UPDATED';
-const EDITOR_HEIGHT_UPDATED = 'javalab/EDITOR_HEIGHT_UPDATED';
-const LEFT_WIDTH_UPDATED = 'javalab/LEFT_WIDTH_UPDATED';
-const RIGHT_WIDTH_UPDATED = 'javalab/RIGHT_WIDTH_UPDATED';
-const SET_INSTRUCTIONS_HEIGHT = 'javalab/SET_INSTRUCTIONS_HEIGHT';
-const SET_INSTRUCTIONS_FULL_HEIGHT = 'javalab/SET_INSTRUCTIONS_FULL_HEIGHT';
 const SET_IS_RUNNING = 'javalab/SET_IS_RUNNING';
 const SET_IS_TESTING = 'javalab/SET_IS_TESTING';
-const SET_CONSOLE_HEIGHT = 'javalab/SET_CONSOLE_HEIGHT';
-const EDITOR_COLUMN_HEIGHT = 'javalab/EDITOR_COLUMN_HEIGHT';
 const SET_BACKPACK_ENABLED = 'javalab/SET_BACKPACK_ENABLED';
 const SET_IS_START_MODE = 'javalab/SET_IS_START_MODE';
 const SET_LEVEL_NAME = 'javalab/SET_LEVEL_NAME';
-const TOGGLE_VISUALIZATION_COLLAPSED = 'javalab/TOGGLE_VISUALIZATION_COLLAPSED';
 const SET_IS_READONLY_WORKSPACE = 'javalab/SET_IS_READONLY_WORKSPACE';
 const SET_HAS_OPEN_CODE_REVIEW = 'javalab/SET_HAS_OPEN_CODE_REVIEW';
 const SET_COMMIT_SAVE_STATUS = 'javalab/SET_COMMIT_SAVE_STATUS';
 const SET_VALIDATION_PASSED = 'javalab/SET_VALIDATION_PASSED';
 const SET_HAS_RUN_OR_TESTED = 'javalab/SET_HAS_RUN_OR_TESTED';
 const SET_IS_JAVABUILDER_CONNECTING = 'javalab/SET_IS_JAVABUILDER_CONNECTING';
-const INCREASE_EDITOR_FONT_SIZE = 'javalab/INCREASE_EDITOR_FONT_SIZE';
-const DECREASE_EDITOR_FONT_SIZE = 'javalab/DECREASE_EDITOR_FONT_SIZE';
 
-// Exported for test
-export const initialState = {
-  displayTheme: DisplayTheme.LIGHT,
-  renderedEditorHeight: 400,
-  leftWidth: 400,
-  rightWidth: 400,
-  instructionsHeight: 200,
-  instructionsFullHeight: 200,
+const initialState = {
   isRunning: false,
   isTesting: false,
-  consoleHeight: 200,
-  editorColumnHeight: 600,
   backpackEnabled: false,
   isStartMode: false,
   levelName: undefined,
-  isVisualizationCollapsed: false,
   isReadOnlyWorkspace: false,
   hasOpenCodeReview: false,
   isCommitSaveInProgress: false,
   hasCommitSaveError: false,
   validationPassed: false,
   hasRunOrTestedCode: false,
-  isJavabuilderConnecting: false,
-  editorFontSize: DEFAULT_FONT_SIZE_PX,
-  canIncreaseFontSize: DEFAULT_FONT_SIZE_PX < MAX_FONT_SIZE_PX,
-  canDecreaseFontSize: DEFAULT_FONT_SIZE_PX > MIN_FONT_SIZE_PX
-};
-
-// Updates the user preferences to reflect change
-export const setDisplayTheme = displayTheme => {
-  new UserPreferences().setDisplayTheme(displayTheme);
-  return {
-    displayTheme: displayTheme,
-    type: COLOR_PREFERENCE_UPDATED
-  };
+  isJavabuilderConnecting: false
 };
 
 export const setIsRunning = isRunning => ({
@@ -80,10 +38,6 @@ export const setIsTesting = isTesting => ({
 export const setBackpackEnabled = backpackEnabled => ({
   type: SET_BACKPACK_ENABLED,
   backpackEnabled
-});
-
-export const toggleVisualizationCollapsed = () => ({
-  type: TOGGLE_VISUALIZATION_COLLAPSED
 });
 
 /**
@@ -105,49 +59,6 @@ export const setLevelName = levelName => ({
 export const setIsJavabuilderConnecting = isJavabuilderConnecting => ({
   type: SET_IS_JAVABUILDER_CONNECTING,
   isJavabuilderConnecting
-});
-
-export const increaseEditorFontSize = () => ({
-  type: INCREASE_EDITOR_FONT_SIZE
-});
-
-export const decreaseEditorFontSize = () => ({
-  type: DECREASE_EDITOR_FONT_SIZE
-});
-
-export const setRenderedHeight = height => ({
-  type: EDITOR_HEIGHT_UPDATED,
-  height
-});
-
-export const setLeftWidth = width => ({
-  type: LEFT_WIDTH_UPDATED,
-  width
-});
-
-export const setRightWidth = width => ({
-  type: RIGHT_WIDTH_UPDATED,
-  width
-});
-
-export const setInstructionsHeight = height => ({
-  type: SET_INSTRUCTIONS_HEIGHT,
-  height
-});
-
-export const setInstructionsFullHeight = height => ({
-  type: SET_INSTRUCTIONS_FULL_HEIGHT,
-  height
-});
-
-export const setConsoleHeight = height => ({
-  type: SET_CONSOLE_HEIGHT,
-  height
-});
-
-export const setEditorColumnHeight = editorColumnHeight => ({
-  type: EDITOR_COLUMN_HEIGHT,
-  editorColumnHeight
 });
 
 export const setIsReadOnlyWorkspace = isReadOnlyWorkspace => ({
@@ -181,42 +92,6 @@ export const setHasRunOrTestedCode = hasRunOrTestedCode => ({
 
 // Reducer
 export default function reducer(state = initialState, action) {
-  if (action.type === COLOR_PREFERENCE_UPDATED) {
-    return {
-      ...state,
-      displayTheme: action.displayTheme
-    };
-  }
-  if (action.type === EDITOR_HEIGHT_UPDATED) {
-    return {
-      ...state,
-      renderedEditorHeight: action.height
-    };
-  }
-  if (action.type === LEFT_WIDTH_UPDATED) {
-    return {
-      ...state,
-      leftWidth: action.width
-    };
-  }
-  if (action.type === RIGHT_WIDTH_UPDATED) {
-    return {
-      ...state,
-      rightWidth: action.width
-    };
-  }
-  if (action.type === SET_INSTRUCTIONS_HEIGHT) {
-    return {
-      ...state,
-      instructionsHeight: action.height
-    };
-  }
-  if (action.type === SET_INSTRUCTIONS_FULL_HEIGHT) {
-    return {
-      ...state,
-      instructionsFullHeight: action.height
-    };
-  }
   if (action.type === SET_IS_RUNNING) {
     return {
       ...state,
@@ -227,18 +102,6 @@ export default function reducer(state = initialState, action) {
     return {
       ...state,
       isTesting: action.isTesting
-    };
-  }
-  if (action.type === SET_CONSOLE_HEIGHT) {
-    return {
-      ...state,
-      consoleHeight: action.height
-    };
-  }
-  if (action.type === EDITOR_COLUMN_HEIGHT) {
-    return {
-      ...state,
-      editorColumnHeight: action.editorColumnHeight
     };
   }
   if (action.type === SET_BACKPACK_ENABLED) {
@@ -257,12 +120,6 @@ export default function reducer(state = initialState, action) {
     return {
       ...state,
       levelName: action.levelName
-    };
-  }
-  if (action.type === TOGGLE_VISUALIZATION_COLLAPSED) {
-    return {
-      ...state,
-      isVisualizationCollapsed: !state.isVisualizationCollapsed
     };
   }
   if (action.type === SET_IS_READONLY_WORKSPACE) {
@@ -300,30 +157,6 @@ export default function reducer(state = initialState, action) {
     return {
       ...state,
       isJavabuilderConnecting: action.isJavabuilderConnecting
-    };
-  }
-  if (action.type === INCREASE_EDITOR_FONT_SIZE) {
-    const newFontSize = Math.min(
-      MAX_FONT_SIZE_PX,
-      state.editorFontSize + FONT_SIZE_INCREMENT_PX
-    );
-    return {
-      ...state,
-      editorFontSize: newFontSize,
-      canIncreaseFontSize: newFontSize < MAX_FONT_SIZE_PX,
-      canDecreaseFontSize: newFontSize > MIN_FONT_SIZE_PX
-    };
-  }
-  if (action.type === DECREASE_EDITOR_FONT_SIZE) {
-    const newFontSize = Math.max(
-      MIN_FONT_SIZE_PX,
-      state.editorFontSize - FONT_SIZE_INCREMENT_PX
-    );
-    return {
-      ...state,
-      editorFontSize: newFontSize,
-      canIncreaseFontSize: newFontSize < MAX_FONT_SIZE_PX,
-      canDecreaseFontSize: newFontSize > MIN_FONT_SIZE_PX
     };
   }
   return state;
